Refactor: (Feedback) 테스트에서만 사용하는 이동전략, 테스트 하위 패키지로 이동
